+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/types/søknadsdata/MedlemskapSøknadsdata.ts
@@ -1,4 +1,4 @@
-import { Utenlandsopphold } from '@navikt/sif-common-forms-ds/src/forms/utenlandsopphold/types';
+import { UtenlandsoppholdEnkel } from '@navikt/sif-common-forms-ds/src/forms/utenlandsopphold/types';
 
 export interface MedlemskapSøknadsdataHarIkkeBoddSkalIkkeBo {
     type: 'harIkkeBoddSkalIkkeBo';
@@ -8,7 +8,7 @@ export interface MedlemskapSøknadsdataHarIkkeBoddSkalIkkeBo {
 export interface MedlemskapSøknadsdataHarBodd {
     type: 'harBodd';
     harBoddUtenforNorgeSiste12Mnd: true;
-    utenlandsoppholdSiste12Mnd: Utenlandsopphold[];
+    utenlandsoppholdSiste12Mnd: UtenlandsoppholdEnkel[];
     skalBoUtenforNorgeNeste12Mnd: false;
 }
 
@@ -16,15 +16,15 @@ export interface MedlemskapSøknadsdataSkalBo {
     type: 'skalBo';
     harBoddUtenforNorgeSiste12Mnd: false;
     skalBoUtenforNorgeNeste12Mnd: true;
-    utenlandsoppholdNeste12Mnd: Utenlandsopphold[];
+    utenlandsoppholdNeste12Mnd: UtenlandsoppholdEnkel[];
 }
 
 export interface MedlemskapSøknadsdataHarBoddSkalBo {
     type: 'harBoddSkalBo';
     harBoddUtenforNorgeSiste12Mnd: true;
-    utenlandsoppholdSiste12Mnd: Utenlandsopphold[];
+    utenlandsoppholdSiste12Mnd: UtenlandsoppholdEnkel[];
     skalBoUtenforNorgeNeste12Mnd: true;
-    utenlandsoppholdNeste12Mnd: Utenlandsopphold[];
+    utenlandsoppholdNeste12Mnd: UtenlandsoppholdEnkel[];
 }
 
 export type MedlemskapSøknadsdata =