+++ apps/pleiepenger-sykt-barn/src/app/søknad/medlemskap-step/MedlemskapStep.stories.tsx
@@ -1,16 +1,16 @@
 import { Meta, StoryObj } from '@storybook/react';
 import { withStepWrapper } from '../../../storybook/decorators';
-import MedlemsskapStep from './MedlemsskapStep';
+import MedlemskapStep from './MedlemskapStep';
 
-const meta: Meta<typeof MedlemsskapStep> = {
-    title: 'Step/MedlemsskapStep',
-    component: MedlemsskapStep,
+const meta: Meta<typeof MedlemskapStep> = {
+    title: 'Step/MedlemskapStep',
+    component: MedlemskapStep,
     decorators: [withStepWrapper],
 };
 
 export default meta;
 
-type Story = StoryObj<typeof MedlemsskapStep>;
+type Story = StoryObj<typeof MedlemskapStep>;
 
 export const Default: Story = {
     args: {},