+++ packages/sif-common-forms-ds/src/forms/medlemskap/summary/MedlemskapSummary.tsx
@@ -1,28 +1,39 @@
 import { FormSummary } from '@navikt/ds-react';
 import { EditStepLink } from '@navikt/sif-common-soknad-ds';
 import { JaNeiSvar, SummaryList } from '@navikt/sif-common-ui';
-import { AppText } from '../../../i18n';
-import { MedlemskapApiData } from '../../../types/søknad-api-data/SøknadApiData';
-import { renderUtenlandsoppholdSummary } from '../summaryItemRenderers';
+import { BostedUtlandApiData, MedlemskapApiData } from '../types';
+import { dateFormatter, ISODateToDate } from '@navikt/sif-common-utils';
+import { MedlemskapFormText } from '../i18n/medlemskapMessages';
 
 export interface Props {
     medlemskap: MedlemskapApiData;
     onEdit?: () => void;
 }
 
+const renderUtenlandsoppholdSummary = (opphold: BostedUtlandApiData): React.ReactNode => (
+    <MedlemskapFormText
+        id="@forms.medlemskapForm.summary.tidsperiode"
+        values={{
+            fraOgMed: dateFormatter.compact(ISODateToDate(opphold.fraOgMed)),
+            tilOgMed: dateFormatter.compact(ISODateToDate(opphold.tilOgMed)),
+            landnavn: opphold.landnavn,
+        }}
+    />
+);
+
 const MedlemskapSummary: React.FC<Props> = ({ medlemskap, onEdit }) => {
     return (
         <FormSummary>
             <FormSummary.Header>
                 <FormSummary.Heading level="2">
-                    <AppText id="medlemskap.summary.header" />
+                    <MedlemskapFormText id="@forms.medlemskapForm.summary.header" />
                 </FormSummary.Heading>
                 {onEdit && <EditStepLink onEdit={onEdit} />}
             </FormSummary.Header>
             <FormSummary.Answers>
                 <FormSummary.Answer>
                     <FormSummary.Label>
-                        <AppText id="steg.oppsummering.utlandetSiste12.header" />
+                        <MedlemskapFormText id="@forms.medlemskapForm.summary.utlandetSiste12.header" />
                     </FormSummary.Label>
                     <FormSummary.Value>
                         <JaNeiSvar harSvartJa={medlemskap.harBoddIUtlandetSiste12Mnd} />
@@ -31,7 +42,7 @@ const MedlemskapSummary: React.FC<Props> = ({ medlemskap, onEdit }) => {
                 {medlemskap.harBoddIUtlandetSiste12Mnd === true && medlemskap.utenlandsoppholdSiste12Mnd.length > 0 && (
                     <FormSummary.Answer>
                         <FormSummary.Label>
-                            <AppText id="steg.oppsummering.utlandetSiste12.liste.header" />
+                            <MedlemskapFormText id="@forms.medlemskapForm.summary.utlandetSiste12.liste.header" />
                         </FormSummary.Label>
                         <FormSummary.Value>
                             <SummaryList
@@ -44,7 +55,7 @@ const MedlemskapSummary: React.FC<Props> = ({ medlemskap, onEdit }) => {
                 )}
                 <FormSummary.Answer>
                     <FormSummary.Label>
-                        <AppText id="steg.oppsummering.utlandetNeste12.header" />
+                        <MedlemskapFormText id="@forms.medlemskapForm.summary.utlandetNeste12.header" />
                     </FormSummary.Label>
                     <FormSummary.Value>
                         <JaNeiSvar harSvartJa={medlemskap.skalBoIUtlandetNeste12Mnd} />
@@ -54,7 +65,7 @@ const MedlemskapSummary: React.FC<Props> = ({ medlemskap, onEdit }) => {
                 {medlemskap.skalBoIUtlandetNeste12Mnd === true && medlemskap.utenlandsoppholdNeste12Mnd.length > 0 && (
                     <FormSummary.Answer>
                         <FormSummary.Label>
-                            <AppText id="steg.oppsummering.utlandetSiste12.liste.header" />
+                            <MedlemskapFormText id="@forms.medlemskapForm.summary.utlandetNeste12.liste.header" />
                         </FormSummary.Label>
                         <FormSummary.Value>
                             <SummaryList