+++ packages/sif-common-forms-ds/src/forms/medlemskap/utils/medlemskapApiUtils.ts
@@ -1,10 +1,10 @@
+import { getCountryName } from '@navikt/sif-common-formik-ds/src/utils/countryUtils';
 import { BostedUtland } from '@navikt/sif-common-forms-ds';
 import { dateToISODate } from '@navikt/sif-common-utils';
-import { getCountryName } from '@navikt/sif-common-formik-ds/src/utils/countryUtils';
-import { MedlemskapSøknadsdata } from '../../types/søknadsdata/MedlemskapSøknadsdata';
-import { BostedUtlandApiData, MedlemskapApiData } from '../../types/søknadApiData/SøknadApiData';
+import { BostedUtlandApiData, MedlemskapApiData } from '../types';
+import { MedlemskapSøknadsdata } from '../types/MedlemskapSøknadsdata';
 
-export const mapBostedUtlandToApiData = (opphold: BostedUtland, locale: string): BostedUtlandApiData => ({
+const mapBostedUtlandToApiData = (opphold: BostedUtland, locale: string): BostedUtlandApiData => ({
     landnavn: getCountryName(opphold.landkode, locale),
     landkode: opphold.landkode,
     fraOgMed: dateToISODate(opphold.fom),